Move LocalDataStore to its own file
